requirements.txt
alembic==1.6.5
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astral==2.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.13.3
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
infinity==1.5
intervals==0.9.2
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.5.1
langdetect==1.0.9
Mako==1.1.4
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
rq==1.9.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.20
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.6
validators==0.19.0
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
WTForms==2.3.3
WTForms-Components==0.10.5
